Make member portals more precise.
Previously, we would consider writes of a property `p` on an object that flows into the entry of a base portal `b` as writes to `(member p b)`.

However, this becomes extremely imprecise if there is a function into which many otherwise unrelated objects are passed. A typical example of this is `util.inherits`: with the previous approach, we would start conflating like-named properties of all classes that are passed as the first argument to that function.