JavaScript: Introduce nodes modelling default `module` and `exports`.
This PR introduces for each CommonJS module two data-flow nodes that represent the initial values of `module` and `exports`, respectively, as well as a node representing the initialization of `module.exports` to `exports`. Moreover, data-flow edges are introduced from the `module` and `exports` nodes to all (non-overwritten) references to `module` and `exports`.